% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_season_stats.R
\name{full_season_stats}
\alias{full_season_stats}
\title{Generate Full Season Stats for a Team}
\usage{
full_season_stats(x, teamName, resultColumn = FTR, resultPrefix = FT)
}
\arguments{
\item{x}{A data frame of match data.}

\item{teamName}{The name of the team to calculate stats for.}

\item{resultColumn}{The column representing the match result.}

\item{resultPrefix}{Prefix for home and away goal columns (default is "FT" for full-time).}
}
\value{
A data frame with the team's season stats.
}
\description{
This function calculates the full season statistics for a given team in the English Premier League dataset.
}
